+++ src/Components/Main/Main.jsx
@@ -1,34 +1,38 @@
-import React,{useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 import Carrousel from '../Carrousel/Carrousel';
 import CardsEventos from '../CardsEventos/CardsEventos'
 // css
 import './Main.css'
 
 const Main = () => {
-  // const [eventos, setEventos] = useState([])
-  // const [isLoading, setIsLoading] = useState(false)
+  const [eventos, setEventos] = useState([])
+  const [isLoading, setIsLoading] = useState(false)
 
-  // const getEventos = async () => {
-  //   try {
-  //     const response = await fetch('http://localhost:5000/eventos')
-  //     const eventosdb = await response.json();
-  //     setEventos(eventosdb)
-  //     setIsLoading(!isLoading)
+  const getEventos = async () => {
+    try {
+      const response = await fetch('http://localhost:5000/eventos')
+      const eventosdb = await response.json();
+      setEventos(eventosdb)
+      setIsLoading(!isLoading)
 
-  //   } catch (error) {
-  //     console.warn(error)
-  //   }
-  // }
+    } catch (error) {
+      console.warn(error)
+    }
+  }
 
-  // useEffect(() => {
-  //   getEventos();
-  // }, [])
+  useEffect(() => {
+    getEventos();
+  }, [])
 
   return (
-   <>
-   <Carrousel />
-   <CardsEventos />
-   </>
+    <>
+      <Carrousel />
+      <div className='container d-flex justify-content-center'>
+      {
+        isLoading ?  eventos.map(evento =>  <CardsEventos evento={evento} /> ) : <p className='pt-5'>Cargando Eventos...</p>
+      }
+      </div>
+    </>
   )
 }
